chore: ini mutation testnet
