+++ apps/app/app/(wallet)/[address]/settings/(components)/settings-testnet-card.tsx
@@ -15,7 +15,6 @@
 
 "use client";
 
-import { updateWalletSettings } from "@lightdotso/client";
 import {
   Button,
   Form,
@@ -28,18 +27,17 @@ import {
   Switch,
 } from "@lightdotso/ui";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { useQueryClient, useMutation } from "@tanstack/react-query";
 import { useState, type FC, useEffect, useMemo } from "react";
 import { useForm } from "react-hook-form";
 import type { Address } from "viem";
 import * as z from "zod";
 import { SettingsCard } from "@/components/settings/settings-card";
 import { TITLES } from "@/const/titles";
-import type { WalletSettingsData } from "@/data";
 import { useDelayedValue } from "@/hooks/useDelayedValue";
-import { useSuspenseQueryWalletSettings } from "@/query";
-import { queryKeys } from "@/queryKeys";
-import { errorToast, successToast } from "@/utils";
+import {
+  useMutationWalletSettings,
+  useSuspenseQueryWalletSettings,
+} from "@/query";
 
 // -----------------------------------------------------------------------------
 // Schema
@@ -77,8 +75,6 @@ export const SettingsTestnetCard: FC<SettingsTestnetCardProps> = ({
   // Query
   // ---------------------------------------------------------------------------
 
-  const queryClient = useQueryClient();
-
   const { walletSettings } = useSuspenseQueryWalletSettings({
     address: address,
   });
@@ -87,78 +83,8 @@ export const SettingsTestnetCard: FC<SettingsTestnetCardProps> = ({
   // Mutate
   // ---------------------------------------------------------------------------
 
-  const { mutate, isPending, isSuccess, isError } = useMutation({
-    mutationFn: async (data: WalletSettingsData) => {
-      const res = await updateWalletSettings({
-        params: {
-          query: {
-            address: address,
-          },
-        },
-        body: {
-          wallet_settings: {
-            is_enabled_testnet: data.is_enabled_testnet,
-          },
-        },
-      });
-
-      // Return if the response is 200
-      res.match(
-        _ => {
-          successToast("Successfully updated wallet settings.");
-        },
-        err => {
-          if (err instanceof Error) {
-            errorToast(err.message);
-          }
-        },
-      );
-    },
-    // When mutate is called:
-    onMutate: async (walletSettings: WalletSettingsData) => {
-      // Cancel any outgoing refetches
-      // (so they don't overwrite our optimistic update)
-      await queryClient.cancelQueries({
-        queryKey: queryKeys.wallet.settings({ address }).queryKey,
-      });
-
-      // Snapshot the previous value
-      const previousSettings: WalletSettingsData | undefined =
-        queryClient.getQueryData(
-          queryKeys.wallet.settings({ address }).queryKey,
-        );
-
-      // Optimistically update to the new value
-      queryClient.setQueryData(
-        queryKeys.wallet.settings({ address }).queryKey,
-        (old: WalletSettingsData) => {
-          return { ...old, walletSettings };
-        },
-      );
-
-      // Return a context object with the snapshotted value
-      return { previousSettings };
-    },
-    // If the mutation fails, use the context we returned above
-    onError: (err, _newWalletSettings, context) => {
-      queryClient.setQueryData(
-        queryKeys.wallet.settings({ address }).queryKey,
-        context?.previousSettings,
-      );
-
-      if (err instanceof Error) {
-        errorToast(err.message);
-      }
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({
-        queryKey: queryKeys.wallet.settings({ address }).queryKey,
-      });
-
-      // Invalidate the cache for the address
-      fetch(`/api/revalidate/tag?tag=${address}`);
-    },
-    mutationKey: queryKeys.wallet.settings({ address }).queryKey,
+  const { mutate, isSuccess, isError, isPending } = useMutationWalletSettings({
+    address: address,
   });
 
   // ---------------------------------------------------------------------------
